CPU build for project 5
Really makes you think about talking in binary, and being able to understand based on what number instruction given, and in exchange being able to do that task lightning fast. Truly lucky to stand on the shoulders of giants; because we don't have to always tinker around at this level as a programmer.