Project 6 Assembler
Created an Assembler in Python, which includes my own test and results indicating that it is true.. If you wish to test yourself simply delete the results and run the test program to create a new file with the tested results. Also I included some history of assemblers and some general improvements over time, and some information on Microsofts MASM assembler. I explain as well what each portion of the assembler is doing, such a striping and cleaning directions given, and handling labels like LOOP.